+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sundaeswap/cli
 
+## 1.1.3
+
+### Patch Changes
+
+- eb04f24: Updates dependencies.
+- Updated dependencies [eb04f24]
+  - @sundaeswap/core@2.6.5
+
+## 1.1.2
+
+### Patch Changes
+
+- 9448981: Updates class references when adding referenceInputs to Blaze's transaction builder. This is an attempt to avoid private class member access somewhere in the tree.
+- Updated dependencies [9448981]
+  - @sundaeswap/core@2.6.3
+
 ## 1.1.1
 
 ### Patch Changes