+++ packages/gummiworm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2.0.1
+
+### Patch Changes
+
+- eb04f24: Updates dependencies.
+
 ## 2.0.0
 
 ### Patch Changes